Regenerate patient-list measure reports for R4 measures
- Previously the measure report calculations included patients we
didn't mean to include. There were a few patients that snuck in via
the measure bundles from the connectathon repo, but those bundles
no longer contain those patients, so I just recalculated with
all the same synthea patients as before.
- Previously CMS130 didn't even have an R4 measure report, so now
it does
- Since the time we originally genereated these patients,
fhir-bundle-calculator has gained the ability to generate individual
measure reports in addition to the patient-list measure report, so
I added all of those to the repo as well